update controllers and model, add schema

+++ controllers/shoppingList/deleteProductFromShoppingList.js
@@ -1,20 +1,17 @@
-const { Product } = require("../../models/product");
+const  Product  = require("../../models/product");
 const { HttpError } = require("../../helpers");
 
 const deleteProductFromShoppingList = async (req, res) => {
-        const { _id: userId } = req.user;
-        const { id: productId} = req.params;
+  const { _id: owner } = req.user;
+  const { id: _id } = req.params;
 
-        const result = await Product.findByIdAndRemove (
-                { user: userId },
-                { _id: productId}
-        );
+  const result = await Product.findByIdAndRemove({ _id, owner   });
 
-        if (!result) {
-                throw HttpError(404, "Product not found in Shopping List");
-        }
+  if (!result) {
+    throw HttpError(404, "Product not found in Shopping List");
+  }
 
-        res.json({message: "Product delete successfully from Shopping List"});
+  res.json({ message: "Product delete successfully from Shopping List" });
 };
 
-module.exports = deleteProductFromShoppingList ;
+module.exports = deleteProductFromShoppingList;
